AshAuthentication.Phoenix.Components.Banner (ash_authentication_phoenix v2.1.1)

Renders a very simple banner at the top of the sign-in component.

Can show either an image or some text, depending on the provided overrides.

Props

overrides - A list of override modules.

Overrides

This component provides the following overrides:

:text - Banner text. Set to nil to disable.

:root_class - CSS class for the root div element.

:href_class - CSS class for the a tag.

:href_url - A URL for the banner image to link to. Set to nil to disable.

:image_class - CSS class for the img tag.

:dark_image_class - Css class for the img tag in dark mode.

:image_url - A URL for the img src attribute. Set to nil to disable.

:dark_image_url - A URL for the img src attribute in dark mode. Set to nil to disable.

:text_class - CSS class for the text div.

See AshAuthentication.Phoenix.Overrides for more information.

Types

props()

@type props() :: %{optional(:overrides) => [module()]}